New Files description (25 Aug 2024)
New final tube rocket is the file that is used to 3D print the tube of the rocket. (The 3D printed tube will be used to make the final body of the rocket using composite process and fiber glass). The .3mf file is the above mentioned file's 3D Printing version. This file has inner diameter of tube as 65 mm and outer diameter of tube as 67 mm. It is very thin as we don't have much PVA filament to 3D print.

The edited rocket body tube for 3Dprint file has inner diameter as
 67 mm and outer diameter as 75 mm.
we need the outer diameter as 67 ( As the file mentioned in the paragraph above).So this file wasn't used to 3D print the body.

(25 Aug 2024)
Mithil Jha